refactor(adapters): close adapter parity checkpoint
Why:
- finish Area 2 with behavior-level adapter parity tied to the final refactored package layout
- replace the remaining oversized Binance and Coinbase adapter files with a standardized package split plus provider-specific extension seams
- restore missing Binance parity behaviors around cutoff-hint suppression and timezone ambiguity blocking

What:
- split Binance into matching, timestamp, field parsing, funding, spot, order, and transaction-history modules under the provider package
- split Coinbase into matching, retail CSV, retail row, timestamp, and asset-migration modules under the provider package
- move Binance and Coinbase unit coverage into provider-local test packages and add direct service proof for normalization-time timezone blocking and stale profile replacement
- update the master parity ledger to the final adapter proof paths and mark removed legacy exception-decision/cache surfaces as superseded-direct

Checks:
- uv run pytest tests/unit/adapters/binance tests/unit/adapters/coinbase tests/unit/adapters/test_adapter_matching_parity.py tests/unit/adapters/test_wallet_inventory_parity.py tests/unit/application/services/test_normalization_service.py tests/contract/test_source_adapter_packs.py tests/contract/test_source_pack_profiles.py --no-cov -q
- uv run pytest tests/unit/test_master_parity_ledger.py tests/unit/test_docs_runtime_parity.py --no-cov -q
- uv run pre-commit run markdownlint --all-files
- uv run python -m tools.run_quality_gates --full-tests
