+++ tests/unit/adapters/binance/test_binance_normalize.py
@@ -1,0 +1,117 @@
+from __future__ import annotations
+
+from pathlib import Path
+
+from crypto_reconciliation.adapters.sources.binance.adapter import BinanceAdapter
+from tests.support.services import build_source_profile
+
+
+def test_binance_adapter_handles_supported_and_review_required_rows(tmp_path: Path) -> None:
+    raw_dir = tmp_path
+    (raw_dir / "Binance-Spot-Trade-History-202603230406(UTC--6)_abcd.csv").write_text(
+        "Time,Pair,Side,Price,Executed,Amount,Fee\n"
+        "23-09-20 18:20:55,ALGOUSDT,SELL,0.0997,103ALGO,10.2691USDT,0.00003593BNB\n",
+        encoding="utf-8",
+    )
+    (raw_dir / "Binance-Deposit-History-202603230411(UTC--6)_abcd.csv").write_text(
+        "Time,Coin,Network,Amount,Address,TXID,Status\n23-05-06 23:05:55,USDT,MATIC,125.564991,addr,tx-dep,Completed\n",
+        encoding="utf-8",
+    )
+    (raw_dir / "Binance-Withdraw-History-202603230412(UTC--6)_abcd.csv").write_text(
+        "Time,Coin,Network,Amount,Fee,Address,TXID,Status\n"
+        "23-09-20 22:25:57,HNT,SOL,12.013,0.11,addr,tx-wd,Completed\n",
+        encoding="utf-8",
+    )
+    (raw_dir / "Binance-Transaction-History-202603230400(UTC--6)_abcd.csv").write_text(
+        "User ID,Time,Account,Operation,Coin,Change,Remark\n"
+        "1,23-08-06 02:34:03,Spot,ETH 2.0 Staking Rewards,BETH,0.00000599,\n"
+        "1,23-09-20 18:46:46,Spot,Small Assets Exchange BNB,ETH,-0.00005643,ETH to BNB\n"
+        "1,23-09-20 18:46:46,Spot,Small Assets Exchange BNB,BNB,0.00041767,ETH to BNB\n"
+        "1,23-09-20 18:17:41,USD-M Futures,Transfer Between Spot Account and UM Futures Account,USDT,-43.90477684,\n"
+        "1,23-09-20 18:17:41,Spot,Transfer Between Spot Account and UM Futures Account,USDT,43.90477684,\n"
+        "1,21-05-11 00:44:33,Spot,Binance Convert,ETH,0.03158115,\n",
+        encoding="utf-8",
+    )
+
+    result = BinanceAdapter().normalize(
+        build_source_profile(adapter_id="binance", raw_dir=str(raw_dir), source="Binance"),
+        raw_dir,
+    )
+
+    assert len(result.canonical_events) == 5
+    assert len(result.issues) == 3
+    kinds = [row.event_kind for row in result.canonical_events]
+    assert "Trade" in kinds
+    assert "Deposit" in kinds
+    assert "Withdrawal" in kinds
+    assert "Staking" in kinds
+    assert any("Transfer Between Spot Account and UM Futures Account" in row.message for row in result.issues)
+
+
+def test_binance_convert_date_updated_covers_transaction_history_one_second_skew(tmp_path: Path) -> None:
+    raw_dir = tmp_path
+    (raw_dir / "Binance-Convert-Order-History-202603230441(UTC--6)_abcd.csv").write_text(
+        "Time,Wallet,Pair,Type,Sell,Buy,Price,Inverse Price,Date Updated,Status\n"
+        "21-05-11 00:44:32,SPOT,ETHBUSD,Instant,124.60184573 BUSD,0.03158115 ETH,x,x,21-05-11 00:44:33,Successful\n",
+        encoding="utf-8",
+    )
+    (raw_dir / "Binance-Transaction-History-202603230400(UTC--6)_abcd.csv").write_text(
+        "User ID,Time,Account,Operation,Coin,Change,Remark\n1,21-05-11 00:44:33,Spot,Binance Convert,ETH,0.03158115,\n",
+        encoding="utf-8",
+    )
+
+    result = BinanceAdapter().normalize(
+        build_source_profile(adapter_id="binance", raw_dir=str(raw_dir), source="Binance"),
+        raw_dir,
+    )
+
+    assert len(result.canonical_events) == 1
+    assert len(result.issues) == 0
+
+
+def test_binance_transaction_history_skips_p2p_rows_when_c2c_history_exists(tmp_path: Path) -> None:
+    raw_dir = tmp_path
+    (raw_dir / "Binance-C2C-Order-History-202603230441(UTC--6)_abcd.csv").write_text(
+        "Order Number,Created Time,Order Type,Asset,Quantity,Total Price,Fiat Type,Counterparty,Status\n"
+        "123,23-09-20 19:48:03,SELL,USDT,891,891,CAD,merchant,Completed\n",
+        encoding="utf-8",
+    )
+    (raw_dir / "Binance-Transaction-History-202603230400(UTC--6)_abcd.csv").write_text(
+        "User ID,Time,Account,Operation,Coin,Change,Remark\n"
+        "1,23-09-20 19:48:03,Funding,P2P Trading,USDT,-891,P2P - 123\n",
+        encoding="utf-8",
+    )
+
+    result = BinanceAdapter().normalize(
+        build_source_profile(adapter_id="binance", raw_dir=str(raw_dir), source="Binance"),
+        raw_dir,
+    )
+
+    assert len(result.canonical_events) == 1
+    assert result.canonical_events[0].event_kind == "Trade"
+    assert len(result.issues) == 0
+
+
+def test_binance_adapter_reads_nested_bundle_paths(tmp_path: Path) -> None:
+    raw_dir = tmp_path / "raw"
+    bundle_dir = raw_dir / "bundle_a"
+    bundle_dir.mkdir(parents=True)
+    (bundle_dir / "Binance-C2C-Order-History-202603230441(UTC--6)_abcd.csv").write_text(
+        "Order Number,Created Time,Order Type,Asset,Quantity,Total Price,Fiat Type,Counterparty,Status\n"
+        "123,23-09-20 19:48:03,SELL,USDT,891,891,CAD,merchant,Completed\n",
+        encoding="utf-8",
+    )
+    (bundle_dir / "Binance-Transaction-History-202603230400(UTC--6)_abcd.csv").write_text(
+        "User ID,Time,Account,Operation,Coin,Change,Remark\n"
+        "1,23-09-20 19:48:03,Funding,P2P Trading,USDT,-891,P2P - 123\n",
+        encoding="utf-8",
+    )
+
+    result = BinanceAdapter().normalize(
+        build_source_profile(adapter_id="binance", raw_dir=str(raw_dir), source="Binance"),
+        raw_dir,
+    )
+
+    assert len(result.canonical_events) == 1
+    assert result.canonical_events[0].event_kind == "Trade"
+    assert not result.issues
